Rename files according react guidelines
