+++ src/components/atoms/button/index.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 
-import classes from './Button.css'
+import classes from './styles.css'
 
 const button = props => <button className={classes.Button}>{props.label}</button>
